Creating Forms for backend
I used Java programming language to create form to learn the method that form creator applications uses. This is the main idea of Bahrainouna that I will be implementing later as a website. There will be some restrictions that the form created must follow the theme of Bahrain, but the implementation that I applied on Java programming languages does not have any restrictions. I used the code of my previous project to help me implement code in this project.