Modifications applied to the website
I made the changes to the website. I divided my code into multiple file. The last commit I did in GitHub is to test a new experiment in my code.
It turns out to be that most of my code were not executable that crashed out my website when testing it.
The prove of this crashing website is found in my social media in the following links:
- https://www.instagram.com/reel/DFgQkzfKzCH/?utm_source=ig_web_copy_link&igsh=MzRlODBiNWFlZA==
- https://x.com/AnotherAqwe/status/1885431096881349086
- https://www.tiktok.com/@aqweandalihassan/video/7466187136354766088?is_from_webapp=1&sender_device=pc&web_id=7452443819373823489